refactor(k6): move deploy correctness tests to k6/correctness/
All five k6 specs run by the deploy workflow are moved to k6/correctness/.
Shared utilities (check.ts, litApiServer.ts) remain in k6/ with import
paths updated to ../. GHA workflows updated to match new paths.

new-account.spec.ts: cap vus 10→2, iterations 10→4 and update
http_reqs threshold to count>=4 to fit deploy correctness constraints.
