+++ .github/workflows/k6-integration.yml
@@ -29,4 +29,4 @@ jobs:
       - name: integration.spec.ts
         env:
           BASE_URL: ${{ inputs.api_root_url }}
-        run: k6 run k6/integration.spec.ts
+        run: k6 run k6/correctness/integration.spec.ts
