+++ .github/workflows/k6-lit-action-ecdsa-sign.yml
@@ -29,4 +29,4 @@ jobs:
       - name: lit-action-ecdsa-sign.spec.ts
         env:
           BASE_URL: ${{ inputs.api_root_url }}
-        run: k6 run k6/lit-action-ecdsa-sign.spec.ts
+        run: k6 run k6/correctness/lit-action-ecdsa-sign.spec.ts
