+++ .github/workflows/k6-new-account.yml
@@ -29,4 +29,4 @@ jobs:
       - name: new-account.spec.ts
         env:
           BASE_URL: ${{ inputs.api_root_url }}
-        run: k6 run k6/new-account.spec.ts
+        run: k6 run k6/correctness/new-account.spec.ts
