+++ .github/workflows/k6-smoke.yml
@@ -29,4 +29,4 @@ jobs:
       - name: smoke.spec.ts
         env:
           BASE_URL: ${{ inputs.api_root_url }}
-        run: k6 run k6/smoke.spec.ts
+        run: k6 run k6/correctness/smoke.spec.ts
