+++ k6/correctness/lit-action-ecdsa-sign.spec.ts
@@ -11,8 +11,8 @@
  *   BASE_URL=https://your-instance/core/v1 k6 run k6/lit-action-ecdsa-sign.spec.ts
  */
 import type { Response } from "k6/http";
-import { checkAndLog } from "./check.ts";
-import { LitApiServerClient } from "./litApiServer.ts";
+import { checkAndLog } from "../check.ts";
+import { LitApiServerClient } from "../litApiServer.ts";
 
 const BASE_URL =
   __ENV.BASE_URL ||
